[CODE HEALTH] Fix clang-tidy narrowing-conversions warnings in tests
Apply static_cast at the narrowing site in three test files:

- otlp_metrics_serialization_test.cc: cast loop index when calling
  protobuf data_points(int) accessors (3 sites).
- aggregation_test.cc: cast size_t max_buckets_ on ternary branches
  where the lvalue is declared int (2 sites).
- metric_test_stress.cc: cast hardware_concurrency() result and the
  uint64_t bucket counts/totals used for int64_t accumulation (3 sites).

Decrement clang-tidy warning limits by 8 to match the removed unique
warnings.

Fixes #3978

+++ exporters/otlp/test/otlp_metrics_serialization_test.cc
@@ -278,7 +278,7 @@ TEST(OtlpMetricSerializationTest, Counter)
   EXPECT_EQ(sum.is_monotonic(), true);
   for (size_t i = 0; i < 1; i++)
   {
-    const auto &proto_number_point = sum.data_points(i);
+    const auto &proto_number_point = sum.data_points(static_cast<int>(i));
     EXPECT_EQ(proto_number_point.as_double(), i == 0 ? 10.2 : 20.2);
   }
 
@@ -308,7 +308,7 @@ TEST(OtlpMetricSerializationTest, Histogram)
             proto::metrics::v1::AggregationTemporality::AGGREGATION_TEMPORALITY_CUMULATIVE);
   for (size_t i = 0; i < 1; i++)
   {
-    const auto &proto_number_point = histogram.data_points(i);
+    const auto &proto_number_point = histogram.data_points(static_cast<int>(i));
     EXPECT_EQ(proto_number_point.sum(), i == 0 ? 100.2 : 200.2);
   }
 
@@ -381,7 +381,7 @@ TEST(OtlpMetricSerializationTest, ObservableGauge)
   otlp_exporter::OtlpMetricUtils::ConvertGaugeMetric(data, &gauge);
   for (size_t i = 0; i < 1; i++)
   {
-    const auto &proto_number_point = gauge.data_points(i);
+    const auto &proto_number_point = gauge.data_points(static_cast<int>(i));
     EXPECT_EQ(proto_number_point.as_double(), i == 0 ? 30.2 : 50.2);
   }
